Split S3 and DynamoDB storage code into two Store objects
Following the same pattern as Bigtable column key
